requirements.txt
paddlepaddle>=2.5.0
paddleocr>=2.7.0
mss>=9.0.1
pillow>=10.0.0
pyautogui>=0.9.54
requests>=2.31.0
